[feat]: allow doctor to add patients to a program
Summary:
This commit will add a feature to allow the doctor to add a patient to
a program. Also, filter viable patients to include only patients who are
not already in the program and those that are active.
Finally, fix the Authentication issue in deleting a program from the
dashboard. (change Berer to Bearer in the Auth header)